requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
kafka-python
redis
pyspark